chore(drawer): move files to internal directory
