+++ src/components/drawer/index.js
@@ -1,2 +1,2 @@
-export { default } from './drawer.component';
-export { SidebarContext } from './drawer.component';
+export { default } from './__internal__/drawer.component';
+export { SidebarContext } from './__internal__/drawer.component';
